moving into the component
